libc: transliterate 27 commits of platform mod.rs files + ioctl-decode helpers

+++ src/commonMain/kotlin/io/github/kotlinmania/libc/unix/bsd/freebsdlike/freebsd/freebsd12/X8664.kt
@@ -7,4 +7,4 @@ public const val PROC_KPTI_CTL: CInt = PROC_PROCCTL_MD_MIN
 public const val PROC_KPTI_CTL_ENABLE_ON_EXEC: CInt = 1
 public const val PROC_KPTI_CTL_DISABLE_ON_EXEC: CInt = 2
 public const val PROC_KPTI_STATUS: CInt = PROC_PROCCTL_MD_MIN + 1
-public const val PROC_KPTI_STATUS_ACTIVE: CInt = 0x80000000.toInt()
+public const val PROC_KPTI_STATUS_ACTIVE: CInt = 0x80000000
